move celery client where it is used --> activity manager
