+++ services/web/server/tests/unit/with_dbs/04/test_computation_utils.py
@@ -6,7 +6,7 @@
 import pytest
 from models_library.projects_state import RunningState
 from simcore_postgres_database.models.comp_pipeline import StateType
-from simcore_service_webserver.computation_api import convert_state_from_db
+from simcore_service_webserver.computation_utils import convert_state_from_db
 
 
 @pytest.mark.parametrize(
